// src/repository/openapi/model/dashboard-table-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Trust Score Dashboard API
 * 기업별 신뢰 점수 시계열 데이터 API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DashboardFilters } from './dashboard-filters';
// May contain unused imports in some cases
// @ts-ignore
import type { DateRange } from './date-range';
// May contain unused imports in some cases
// @ts-ignore
import type { PaginationConfig } from './pagination-config';
// May contain unused imports in some cases
// @ts-ignore
import type { SearchFilter } from './search-filter';
// May contain unused imports in some cases
// @ts-ignore
import type { SortConfig } from './sort-config';

/**
 * 대시보드 테이블 요청
 * @export
 * @interface DashboardTableRequest
 */
export interface DashboardTableRequest {
    /**
     * 
     * @type {DateRange}
     * @memberof DashboardTableRequest
     */
    'dateRange': DateRange;
    /**
     * 
     * @type {SearchFilter}
     * @memberof DashboardTableRequest
     */
    'search'?: SearchFilter | null;
    /**
     * 
     * @type {DashboardFilters}
     * @memberof DashboardTableRequest
     */
    'filters'?: DashboardFilters | null;
    /**
     * 
     * @type {SortConfig}
     * @memberof DashboardTableRequest
     */
    'sort'?: SortConfig | null;
    /**
     * 
     * @type {PaginationConfig}
     * @memberof DashboardTableRequest
     */
    'pagination'?: PaginationConfig | null;
}