// src/repository/openapi/model/dashboard-table-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Trust Score Dashboard API
 * 기업별 신뢰 점수 시계열 데이터 API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DashboardTableRow } from './dashboard-table-row';

/**
 * 대시보드 테이블 응답
 * @export
 * @interface DashboardTableResponse
 */
export interface DashboardTableResponse {
    /**
     * 
     * @type {Array<DashboardTableRow>}
     * @memberof DashboardTableResponse
     */
    'rows': Array<DashboardTableRow>;
    /**
     * 
     * @type {number}
     * @memberof DashboardTableResponse
     */
    'total': number;
    /**
     * 
     * @type {number}
     * @memberof DashboardTableResponse
     */
    'currentPage': number;
    /**
     * 
     * @type {number}
     * @memberof DashboardTableResponse
     */
    'pageSize': number;
}